Move configuration to fab.
